Configure CI to run integration tests separately
